# backend/docs/openapi/stashbox-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stashbox API",
    "version": "0.1.0",
    "description": "听匣（Stashbox）后端 4 服务合并 OpenAPI schema。\n\n由 `backend/scripts/export_openapi.py` 自动生成，请勿手改。\n\n服务分组见 tags：api-gateway / user-service / content-service / ai-service。"
  },
...
  "paths": {
...
    "/api/v1/admin/tts/voices": {
      "get": {
        "summary": "Admin List Voices",
        "operationId": "admin_list_voices_api_v1_admin_tts_voices_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "content-service",
          "api-gateway"
        ]
      },
      "post": {
        "summary": "Admin Create Voice",
        "operationId": "admin_create_voice_api_v1_admin_tts_voices_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceCreateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "content-service",
          "api-gateway"
        ]
      }
    },
    "/api/v1/admin/tts/voices/import-from-config": {
      "post": {
        "summary": "Admin Import Voice From Config",
        "description": "把当前全局 TTS 配置里的 indextts_ref_audio 收编成音色库第一条。\n\n存在的理由：迁移 0033 **刻意没有 seed 默认音色** —— 迁移里读 env 会把本机\n绝对路径固化进库，换台机器就是死路径。管理员点一次这个按钮，就能在不敲\n任何路径的前提下完成冷启动。",
        "operationId": "admin_import_voice_from_config_api_v1_admin_tts_voices_import_from_config_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportFromConfigOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "content-service",
          "api-gateway"
        ]
      }
    },
    "/api/v1/admin/tts/voices/{voice_id}": {
      "put": {
        "summary": "Admin Update Voice",
        "operationId": "admin_update_voice_api_v1_admin_tts_voices__voice_id__put",
        "parameters": [
          {
            "name": "voice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Voice Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "content-service",
          "api-gateway"
        ]
      },
      "delete": {
        "summary": "Admin Delete Voice",
        "operationId": "admin_delete_voice_api_v1_admin_tts_voices__voice_id__delete",
        "parameters": [
          {
            "name": "voice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Voice Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "content-service",
          "api-gateway"
        ]
      }
    },
    "/api/v1/admin/tts/voices/{voice_id}/preview": {
      "post": {
        "summary": "Admin Preview Voice",
        "description": "试听：拿这个音色的 ref_audio 真合成一段。\n\n真调 oMLX 而不是假装成功 —— 音色库里最容易翻车的是「参考音频和参考文本\n对不上」「音频根本不是有效 wav」，这两个只有真合成一次才会暴露。",
        "operationId": "admin_preview_voice_api_v1_admin_tts_voices__voice_id__preview_post",
        "parameters": [
          {
            "name": "voice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Voice Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoicePreviewIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoicePreviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "content-service",
          "api-gateway"
        ]
      }
    },
...
    "/api/v1/tts/voices": {
      "get": {
        "summary": "List Available Voices",
        "description": "可选购音色列表（只含已上架的）。\n\n连同 `available_speeds` 一起给：语速档位是服务端单一事实源，\nApp 不该硬编码 —— 后台改了这里就改了全端行为。",
        "operationId": "list_available_voices_api_v1_tts_voices_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "content-service",
          "api-gateway"
        ]
      }
    },
...
    "/api/v1/users/me/tts-preference": {
      "get": {
        "summary": "Get My Tts Preference",
        "description": "我选的音色 + 我的语速。\n\n同时回 `effective_voice_name` / `effective_source`：用户可能从没设过偏好，\n这时实际生效的是全局默认音色或全局配置 —— App 应该显示**实际生效的那个**，\n否则用户会看到「未选择」但听到的是某个音色，典型的假闭环。\n\n⚠️ **已下架 / 已软删的音色要当没选**（2026-09-30 自测发现的假闭环）：\n管理员把用户选中的音色下架后，`resolve_voice_for_user` 会跳过它回落到\n默认/全局配置，但本端点原先照样回它的 `display_name` ——\nApp 于是显示「待下架音色」，可实际生效的根本不是它，而且它也不在\n`GET /api/v1/tts/voices` 的可选列表里，Sheet 中没有对应行可点。\n用户看到的是一个自己选不了、也没在生效的音色，音频还和显示的不一样。\n现在 `voice_id` / `voice_name` 一起置空，如实回落到 `effective_*`。",
        "operationId": "get_my_tts_preference_api_v1_users_me_tts_preference_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TTSPreferenceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "content-service",
          "api-gateway"
        ]
      },
      "put": {
        "summary": "Update My Tts Preference",
        "description": "改音色 / 语速。\n\n**不会触发重跑蒸馏**：语速是播放端行为；音色只对之后新剪藏的文章生效，\n存量文章要换音色得走重生成入口（见 re-distill 端点）。",
        "operationId": "update_my_tts_preference_api_v1_users_me_tts_preference_put",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TTSPreferenceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TTSPreferenceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "content-service",
          "api-gateway"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportFromConfigOut": {
        "properties": {
          "voice": {
            "$ref": "#/components/schemas/AdminVoiceOut"
          },
          "created": {
            "type": "boolean",
            "title": "Created",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "voice"
        ],
        "title": "ImportFromConfigOut"
      },
...
      "VoiceCreateIn": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "ref_audio_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Audio Url"
          },
          "ref_audio_b64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Audio B64"
          },
          "ref_text": {
            "type": "string",
            "title": "Ref Text"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_default": {
            "type": "boolean",
            "title": "Is Default",
            "default": false
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "slug",
          "display_name",
          "ref_text"
        ],
        "title": "VoiceCreateIn"
      },
      "VoiceListOut": {
        "properties": {
          "voices": {
            "items": {
              "$ref": "#/components/schemas/VoiceOut"
            },
            "type": "array",
            "title": "Voices"
          },
          "available_speeds": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Available Speeds"
          },
          "default_speed": {
            "type": "number",
            "title": "Default Speed",
            "default": 1.0
          }
        },
        "type": "object",
        "required": [
          "voices",
          "available_speeds"
        ],
        "title": "VoiceListOut"
      },
...
      "VoicePreviewIn": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "default": "这是一段试听文本，用来确认这个音色的效果。"
          }
        },
        "type": "object",
        "title": "VoicePreviewIn"
      },
      "VoicePreviewOut": {
        "properties": {
          "audio_url": {
            "type": "string",
            "title": "Audio Url"
          },
          "bytes_len": {
            "type": "integer",
            "title": "Bytes Len"
          },
          "duration_sec": {
            "type": "integer",
            "title": "Duration Sec"
          },
          "voice_id": {
            "type": "string",
            "title": "Voice Id"
          }
        },
        "type": "object",
        "required": [
          "audio_url",
          "bytes_len",
          "duration_sec",
          "voice_id"
        ],
        "title": "VoicePreviewOut"
      },
      "VoiceUpdateIn": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "ref_audio_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Audio Url"
          },
          "ref_audio_b64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Audio B64"
          },
          "ref_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref Text"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "is_default": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Default"
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order"
          }
        },
        "type": "object",
        "title": "VoiceUpdateIn"
      },
...
    },
    "securitySchemes": {}
  }
}